* account-api:
  Create deposit transaction when create account
  Add TransactionService and implement MakeDeposit function
  Implement create account api

+++ AwesomeBank/AwesomeBankAPI/Repository/MockTransactionRepository.cs
@@ -21,7 +21,7 @@ namespace AwesomeBankAPI.Repository
                 Amount = 200,
                 Description = "Deposit to Account:09382349123, Amount 200 Baht.",
                 ReceiverAccoundId = null,
-                TransactionType = GlobalConfig.TransactionType.DEPOSIT,
+                TransactionType = (int)GlobalConfig.TransactionType.DEPOSIT,
                 IsActive = true,
                 CreatedDate = DateTime.Now
             },
@@ -34,7 +34,7 @@ namespace AwesomeBankAPI.Repository
                 Amount = 500,
                 Description = "Transfer from Account:09382349123 to Account:09382349123, Amount 500 Baht.",
                 ReceiverAccoundId = Guid.Parse("DD3DDE66-38CC-487A-8A6C-82E5928A7D69"),
-                TransactionType = GlobalConfig.TransactionType.TRANSFER,
+                TransactionType = (int)GlobalConfig.TransactionType.TRANSFER,
                 IsActive = true,
                 CreatedDate = DateTime.Now
             },
